[ADD] create employee filter

+++ models/timesheet_line.py
@@ -105,12 +105,13 @@ class AccountAnalyticLine(models.Model):
 
     @api.multi
     def unlink(self):
-        for line in self:
-            agrs = {
-                "worklog_id": line.id_jira,
-                "task_key": line.task_id.key
-            }
-            services.jira_services.JiraServices(request.session["authorization"]).delete_worklog(agrs)
+        if self.env.user.authorization:
+            for line in self:
+                agrs = {
+                    "worklog_id": line.id_jira,
+                    "task_key": line.task_id.key
+                }
+                services.jira_services.JiraServices(request.session["authorization"]).delete_worklog(agrs)
 
         return super(AccountAnalyticLine, self).unlink()
